Sepration of Logic Done in TabContent View
